requirements.txt
semantic-kernel
fastapi
uvicorn
pydantic
pydantic-settings
requests
python-dotenv>=0.19.0
colorama
flask>=2.0.0
click
pytest>=7.0.0
pytest-asyncio
httpx
pytest-cov>=4.0.0
flake8
mypy
autoflake
isort
autopep8
types-requests
types-colorama
openai>=1.0.0
sqlalchemy>=2.0.0
alembic
aiosqlite
greenlet
fastmcp
supervisor